requirements.txt
faust-streaming==0.4.0
googlemaps==4.4.2
SQLAlchemy==1.3.22
Faker==5.5.1
python-dotenv==0.15.0
psycopg2-binary==2.8.6
